Update sketch.js
I stayed with the glowing ball idea, and I tranformed into a class with the function of reducing the glow. Also, intead of other glowing balls I changed for triangles, also created a class for that, with the function of locating the center of the ball. Also, I created an animation to represent the moment in when the ball is hit. Started working in the scenes, still working to have a few more based on what I have now