Update N3
Eliminated double display in ball (just needed to fill the ellipse in the center), added some variables indicated in draw loop. Eliminated the reset function in Enemie class cause it was bugging the sketch. This as a whole is the main scene now I have to repeat it more times changing the variables in draw